Moved test classes to corresponding packages
