--- owncloudData/src/main/java/com/owncloud/android/data/webfinger/datasources/RemoteWebFingerDatasource.kt
@@ -1,36 +1,0 @@
-/**
- * ownCloud Android client application
- *
- * Copyright (C) 2022 ownCloud GmbH.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.owncloud.android.data.webfinger.datasources
-
-import com.owncloud.android.domain.webfinger.model.WebFingerRel
-
-interface RemoteWebFingerDatasource {
-    fun getInstancesFromWebFinger(
-        lookupServer: String,
-        rel: WebFingerRel,
-        resource: String
-    ): List<String>
-
-    fun getInstancesFromAuthenticatedWebFinger(
-        lookupServer: String,
-        rel: WebFingerRel,
-        resource: String,
-        username: String,
-        accessToken: String,
-    ): List<String>
-}
